Move `formatBankValue` function to `utils` directory
